minor type annotation

+++ src/machines.jl
@@ -6,7 +6,7 @@
 Return the current global default value for scientific type checking
 when constructing machines.
 
-   default_scitype_check_level(i)
+   default_scitype_check_level(i::Integer)
 
 Set the global default value for scientific type checking to `i`.
